fintech-ach-app/backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
pytest==8.0.1
pytest-asyncio==0.23.5
httpx==0.26.0
python-multipart==0.0.6
plaid-python==18.1.0
pydantic==2.5.1
pydantic-settings==2.1.0
aiohttp==3.9.3
urllib3==1.26.18
pytest-cov==4.1.0
pytest-mock==3.12.0
coverage==7.3.2
sqlalchemy==2.0.23
alembic==1.13.1
asyncpg==0.29.0
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
redis[hiredis]==5.0.1
greenlet==3.0.1
PyJWT==2.8.0
email-validator==2.1.0.post1
  